// src/model/PokemonData.ts
export interface PokeApiCommonResponse {
    name: string;
    url: string;
}

export interface Ability {
    ability: PokeApiCommonResponse;
    is_hidden: boolean;
    slot: number;
}

export interface GameIndice {
    game_index: number;
    version: PokeApiCommonResponse;
}

export interface VersionGroupDetail {
    level_learned_at: number;
    move_learn_method: PokeApiCommonResponse;
    version_group: PokeApiCommonResponse;
}

export interface Move {
    move: PokeApiCommonResponse;
    version_group_details: VersionGroupDetail[];
}

export interface DreamWorld {
    front_default: string;
    front_female?: any;
}

export interface OfficialArtwork {
    front_default: string;
}

export interface Other {
    dream_world: DreamWorld;
    "official-artwork": OfficialArtwork;
}
// ...
export interface Versions {
    "generation-i": GenerationI;
    "generation-ii": GenerationIi;
    "generation-iii": GenerationIii;
    "generation-iv": GenerationIv;
    "generation-v": GenerationV;
    "generation-vi": GenerationVi;
    "generation-vii": GenerationVii;
    "generation-viii": GenerationViii;
}

export interface Sprites {
    back_default: string;
    back_female?: any;
    back_shiny: string;
    back_shiny_female?: any;
    front_default: string;
    front_female?: any;
    front_shiny: string;
    front_shiny_female?: any;
    other: Other;
    versions: Versions;
}

export interface Stat {
    base_stat: number;
    effort: number;
    stat: PokeApiCommonResponse;
}

export interface Type {
    slot: number;
    type: PokeApiCommonResponse;
}

export interface PokemonData {
    abilities: Ability[];
    base_experience: number;
    forms: PokeApiCommonResponse[];
    game_indices: GameIndice[];
    height: number;
    held_items: any[];
    id: number;
    is_default: boolean;
    location_area_encounters: string;
    moves: Move[];
    name: string;
    order: number;
    past_types: any[];
    species: PokeApiCommonResponse;
    sprites: Sprites;
    stats: Stat[];
    types: Type[];
    weight: number;
}